pyproject.toml
[tool.poetry]
name = "aiohttp-jwt-auth"
version = "0.1.0"
description = "Library for simple JWT authenticate for an aiohttp application"
authors = ["Hound672"]

[tool.poetry.dependencies]
python = "^3.7"
pyjwt = "1.7.1"
aiohttp = "3.5.4"
cryptography = "2.6.1"
pydantic = "0.31.1"

[tool.poetry.dev-dependencies]
sphinx = "1.8.5"
pytest = "4.3.0"
pytest-aiohttp = "0.3.0"
mypy = "0.670"
pytest-cov = "2.6.1"
factory-boy = "2.11.1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
